created the database
created the databse had to download docker to make a container for azure data studio to run and connect it to visulae studio. Also had to add the csv file to the container. which was done in terminal. Next action fix the back end so it connects to the azure rhather then mongosh.